Extract VSA Job creation in its own package
This commit encapsulates the logic that, given a konflux snapshot,
triggers the creation of a kubernetes job that performs the creation
(and upload) of the VSA for that snapshot.

This allows for great flexibility in the choice of the triggering
mechanism: in this project it is integrated in a knative service, but
after this refactor it could easily be replaced by any triggering logic,
such as a plain controller for example.

It exposes a single interface with a single method, making the
integration with the caller very easy.

Assisted by: Claude Code

Ref: https://issues.redhat.com/browse/EC-1561