Remove non relevant acceptance test
The removed test was asserting the internal logic of the VSA creation.
This is already covered by the acceptance tests of the conforma/cli
repository.

The acceptance tests in this repository only need to ensure that the k8s
job is correctly instantiated, and that it calls the 'ec validate image'
with all the expected parameters.

Assisted by: Claude Code

Ref: https://issues.redhat.com/browse/EC-1561